# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rbac/roles": {
      "post": {
        "operationId": "RbacController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a new role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "get": {
        "operationId": "RbacController_findAllRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of roles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get all roles",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
    "/rbac/roles/{id}": {
      "get": {
        "operationId": "RbacController_findRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a role by ID",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "patch": {
        "operationId": "RbacController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "delete": {
        "operationId": "RbacController_deleteRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role deleted."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Delete a role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
    "/rbac/permissions": {
      "post": {
        "operationId": "RbacController_createPermission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permission created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a new permission",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "get": {
        "operationId": "RbacController_findAllPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get all permissions",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
    "/rbac/permissions/{id}": {
      "get": {
        "operationId": "RbacController_findPermissionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a permission by ID",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "patch": {
        "operationId": "RbacController_updatePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a permission",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "delete": {
        "operationId": "RbacController_deletePermission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission deleted."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Permission not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Delete a permission",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
    "/rbac/roles/{roleId}/permissions/{permissionId}": {
      "post": {
        "operationId": "RbacController_assignPermissionToRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission assigned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role or Permission not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Assign a permission to a role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      },
      "delete": {
        "operationId": "RbacController_removePermissionFromRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission removed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role or Permission not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Remove a permission from a role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
    "/rbac/roles/{roleId}/permissions": {
      "get": {
        "operationId": "RbacController_findPermissionsForRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of permissions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Role not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get permissions for a specific role",
        "tags": [
          "RBAC (Roles & Permissions)"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSO Boilerplate API",
    "description": "API documentation for the SSO Boilerplate application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API Key untuk autentikasi akses API"
      }
    },
    "schemas": {
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Content Manager",
            "description": "Name of the role (unique)",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "example": "Role for managing content and articles",
            "description": "Brief description of role responsibilities",
            "maxLength": 255
          }
        },
        "required": [
          "name"
        ]
      },
      "RoleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "admin"
          },
          "description": {
            "type": "object",
            "example": "Administrator role",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Content Editor",
            "description": "Updated role name",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "example": "Updated role description",
            "description": "Updated description for the role",
            "maxLength": 255
          }
        }
      },
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "example": "create",
            "description": "Action that can be performed (e.g., create, read, update, delete, manage)",
            "maxLength": 100
          },
          "subject": {
            "type": "string",
            "example": "article",
            "description": "Resource or subject on which the action is performed (e.g., user, article, product)",
            "maxLength": 100
          }
        },
        "required": [
          "action",
          "subject"
        ]
      },
      "PermissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "subject",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "example": "manage",
            "description": "Updated action value",
            "maxLength": 100
          },
          "subject": {
            "type": "string",
            "example": "content",
            "description": "Updated subject/resource value",
            "maxLength": 100
          }
        }
      },
...
    }
  }
}